Visitor's Management System
First Application Written In Golang

All Employees are registered on the application to enable easy access of details
and notification

Each Guest on every new visit is registered and given a code, which he/she uses
on every visit to pull up user detail.

Notification is sent to every user (By mail or Text message).

The Three  main features include:

-Employees  Feature
-Visitors Feature
-Visit Feature